+++ src/components/spatial/Spatial.js
@@ -37,6 +37,7 @@ export function square(x, y, r) {
  * @prop {number} moleculeRadius
  * @prop {number} cellOpacity The value for `opacity` to pass
  * to the deck.gl cells PolygonLayer.
+ * @prop {number} lineWidthScale Width of the border shown when the opacity of cells are lowered.
  * @prop {object} imageLayerProps
  * @prop {object} imageLayerLoaders
  * @prop {object} cellColors Object mapping cell IDs to colors.
@@ -72,6 +73,7 @@ export default function Spatial(props) {
     moleculeRadius = 10,
     cellOpacity = 1.0,
     moleculesOpacity = 1.0,
+    lineWidthScale = 10,
     areMoleculesOn = true,
     imageLayerProps = {},
     imageLayerLoaders = {},
@@ -196,11 +198,22 @@ export default function Spatial(props) {
   const cellsLayer = useMemo(() => new SelectablePolygonLayer({
     id: CELLS_LAYER_ID,
     backgroundColor: [0, 0, 0],
-    opacity: cellOpacity,
     isSelected: getCellIsSelected,
-    stroked: false,
+    stroked: true,
     getPolygon: getCellPolygon,
-    getColor: getCellColor,
+    updateTriggers: {
+      getFillColor: [cellOpacity],
+    },
+    getFillColor: (cellEntry) => {
+      const color = getCellColor(cellEntry);
+      color[3] = cellOpacity * 255;
+      return color;
+    },
+    getLineColor: (cellEntry) => {
+      const color = getCellColor(cellEntry);
+      color[3] = 255;
+      return color;
+    },
     onClick: (info) => {
       if (tool) {
         // If using a tool, prevent individual cell selection.
@@ -211,9 +224,12 @@ export default function Spatial(props) {
     },
     visible: areCellsOn,
     ...cellLayerDefaultProps(cellsData, updateStatus, updateCellsHover, uuid),
+    getLineWidth: cellOpacity < 0.7 ? 1 : 0,
+    lineWidthScale,
+
   }), [cellsData, updateStatus, updateCellsHover,
     uuid, onCellClick, tool, getCellColor, getCellPolygon, cellOpacity,
-    getCellIsSelected, areCellsOn]);
+    getCellIsSelected, areCellsOn, lineWidthScale]);
 
   const moleculesLayer = useMemo(() => new ScatterplotLayer({
     id: 'molecules-layer',