+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Make `LayerControllerSubscriber` names booleans-questions.
 - Fix reisizing bug.
 - Bump `vitessce-data` to 0.0.30.
+- Changed the "Please wait" modal to a spinner.
 
 ## [0.1.8](https://www.npmjs.com/package/vitessce/v/0.1.8) - 2020-07-02