+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - Add `labelOverride` prop for genes component.
+- Cell boundaries are visible when opacity drops below a cutoff.
 - Add download lists for files to components that display data.
 - Added `onWarn` callback to the `<Vitessce/>` component to allow a consumer app to manage display of warning messages.