+++ CHANGELOG.md
@@ -20,6 +20,7 @@
 - Fix reisizing bug.
 - Bump `vitessce-data` to 0.0.30.
 - Changed the "Please wait" modal to a spinner.
+- Bump `viv` to 0.3.0.
 
 ## [0.1.8](https://www.npmjs.com/package/vitessce/v/0.1.8) - 2020-07-02